requirements.txt
scikit-learn==1.6.1
pandas==2.2.3
pytorch==2.2.0
numpy==1.26.4